+++ example/lib/page/bloc/bloc_paywall_page.dart
@@ -2,10 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:linkfive_purchases_example/bloc/linkfive_active_products_cubit.dart';
 import 'package:linkfive_purchases_example/bloc/linkfive_products_cubit.dart';
-import 'package:linkfive_purchases_example/main.dart';
 import 'package:linkfive_purchases_example/page/bloc/bloc_paywall.dart';
 import 'package:linkfive_purchases_example/page/bloc/bloc_raw.dart';
-import 'package:linkfive_purchases_provider/linkfive_purchases_provider.dart';
 
 class BlocPaywallPage extends Page {
 
@@ -14,8 +12,6 @@ class BlocPaywallPage extends Page {
   final int page;
 
   BlocPaywallPage(this.page) : super(key: ValueKey("BlocPaywallPage")) {
-    LinkFivePurchases.init(MyApp.linkFiveApiKey,
-        env: LinkFiveEnvironment.STAGING);
   }
 
   @override