+++ packages/linkfive_purchases/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.0.0-dev.1
+## 2.0.0-dev.3
 
 Pre-Release of LinkFive 2.0 The changelog is following soon